libzc: move long running test to EXTRACHECK
As reported on Debian mailing list, the last test in the check_reduce
suite (test_can_generate_next_array_from_plaintext) can run for quite
some and could cause the test to fail with a timeout. Make sure this
does not happen by increasing the timeout and by ifdefing the test with
the EXTRACHECK ifdef.